# Conflicts:
#	src/com/apmath/loans/domain/models/ApplicationInterface.kt
#	src/com/apmath/loans/domain/models/loans/LoanCreationData.kt
#	src/com/apmath/loans/domain/services/PaymentService.kt
#	src/com/apmath/loans/domain/services/PaymentServiceInterface.kt
#	src/com/apmath/loans/infrastructure/fetchers/ApplicationsFetcher.kt
#	src/com/apmath/loans/infrastructure/fetchers/CalculationsFetcher.kt
#	src/com/apmath/loans/infrastructure/fetchers/ClientsFetcher.kt
#	src/com/apmath/loans/infrastructure/models/Application.kt

+++ src/com/apmath/loans/domain/models/loans/LoanInitializationInterface.kt
@@ -5,6 +5,6 @@ import com.apmath.loans.domain.data.Money
 interface LoanInitializationInterface {
     val amount: Money
     val term: Int
-    val interest: Int
+    val interest: Float
     val date: String
 }
